Move Feature Flag Data source to feature_flag folder
