+++ tensorboard/webapp/feature_flag/data_source/tb_feature_flag_data_source_types.ts
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 import {Injectable} from '@angular/core';
-import {FeatureFlags} from '../feature_flag/types';
+import {FeatureFlags} from '../types';
 
 @Injectable()
 export abstract class TBFeatureFlagDataSource {
