+++ tensorboard/webapp/feature_flag/effects/BUILD
@@ -10,10 +10,10 @@ tf_ng_module(
     deps = [
         "//tensorboard/webapp/feature_flag:force_svg_data_source",
         "//tensorboard/webapp/feature_flag/actions",
+        "//tensorboard/webapp/feature_flag/data_source:feature_flag",
+        "//tensorboard/webapp/feature_flag/data_source:feature_flag_types",
         "//tensorboard/webapp/feature_flag/store",
         "//tensorboard/webapp/feature_flag/store:types",
-        "//tensorboard/webapp/webapp_data_source:feature_flag",
-        "//tensorboard/webapp/webapp_data_source:feature_flag_types",
         "@npm//@angular/core",
         "@npm//@ngrx/effects",
         "@npm//@ngrx/store",
@@ -35,9 +35,9 @@ tf_ng_module(
         "//tensorboard/webapp/feature_flag:testing",
         "//tensorboard/webapp/feature_flag:types",
         "//tensorboard/webapp/feature_flag/actions",
+        "//tensorboard/webapp/feature_flag/data_source:feature_flag_testing",
         "//tensorboard/webapp/feature_flag/store",
         "//tensorboard/webapp/feature_flag/store:types",
-        "//tensorboard/webapp/webapp_data_source:feature_flag_testing",
         "@npm//@ngrx/effects",
         "@npm//@ngrx/store",
         "@npm//@types/jasmine",
