+++ tensorboard/webapp/feature_flag/effects/feature_flag_effects.ts
@@ -17,8 +17,8 @@ import {Injectable} from '@angular/core';
 import {Actions, createEffect, ofType} from '@ngrx/effects';
 import {Action, createAction, Store} from '@ngrx/store';
 import {combineLatestWith, map} from 'rxjs/operators';
-import {TBFeatureFlagDataSource} from '../../webapp_data_source/tb_feature_flag_data_source_types';
 import {partialFeatureFlagsLoaded} from '../actions/feature_flag_actions';
+import {TBFeatureFlagDataSource} from '../data_source/tb_feature_flag_data_source_types';
 import {ForceSvgDataSource} from '../force_svg_data_source';
 import {getIsAutoDarkModeAllowed} from '../store/feature_flag_selectors';
 import {State} from '../store/feature_flag_types';
