+++ tensorboard/webapp/feature_flag/feature_flag_module.ts
@@ -21,7 +21,7 @@ import {
   PersistentSettingsConfigModule,
   ThemeValue,
 } from '../persistent_settings';
-import {TBFeatureFlagModule} from '../webapp_data_source/tb_feature_flag_module';
+import {TBFeatureFlagModule} from './data_source/tb_feature_flag_module';
 import {FeatureFlagEffects} from './effects/feature_flag_effects';
 import {ForceSvgDataSourceModule} from './force_svg_data_source_module';
 import {reducers} from './store/feature_flag_reducers';
