+++ tensorboard/webapp/routes/BUILD
@@ -41,11 +41,11 @@ tf_ts_library(
         "//tensorboard/webapp/app_routing:deep_link_provider",
         "//tensorboard/webapp/app_routing:route_config",
         "//tensorboard/webapp/app_routing:types",
+        "//tensorboard/webapp/feature_flag/data_source:feature_flag_types",
         "//tensorboard/webapp/metrics:types",
         "//tensorboard/webapp/metrics/data_source:types",
         "//tensorboard/webapp/runs:types",
         "//tensorboard/webapp/tb_wrapper",
-        "//tensorboard/webapp/webapp_data_source:feature_flag_types",
         "@npm//@angular/core",
         "@npm//@ngrx/store",
         "@npm//rxjs",
