+++ tensorboard/webapp/routes/dashboard_deeplink_provider.ts
@@ -20,6 +20,11 @@ import {DeepLinkProvider} from '../app_routing/deep_link_provider';
 import {SerializableQueryParams} from '../app_routing/types';
 import {State} from '../app_state';
 import {
+  ENABLE_COLOR_GROUP_BY_REGEX_QUERY_PARAM_KEY,
+  ENABLE_COLOR_GROUP_QUERY_PARAM_KEY,
+  EXPERIMENTAL_PLUGIN_QUERY_PARAM_KEY,
+} from '../feature_flag/data_source/tb_feature_flag_data_source_types';
+import {
   isPluginType,
   isSampledPlugin,
   isSingleRunPlugin,
@@ -27,11 +32,6 @@ import {
 import {CardUniqueInfo} from '../metrics/types';
 import {GroupBy, GroupByKey} from '../runs/types';
 import * as selectors from '../selectors';
-import {
-  ENABLE_COLOR_GROUP_BY_REGEX_QUERY_PARAM_KEY,
-  ENABLE_COLOR_GROUP_QUERY_PARAM_KEY,
-  EXPERIMENTAL_PLUGIN_QUERY_PARAM_KEY,
-} from '../webapp_data_source/tb_feature_flag_data_source_types';
 import {
   DeserializedState,
   PINNED_CARDS_KEY,
