+++ tensorboard/webapp/webapp_data_source/BUILD
@@ -62,7 +62,6 @@ tf_ng_module(
         "tb_http_client_test.ts",
     ],
     deps = [
-        ":feature_flag_testing",
         ":http_client",
         ":http_client_testing",
         "//tensorboard/webapp/angular:expect_angular_common_http",
@@ -70,65 +69,10 @@ tf_ng_module(
         "//tensorboard/webapp/angular:expect_ngrx_store_testing",
         "//tensorboard/webapp/app_routing:app_root",
         "//tensorboard/webapp/core:types",
+        "//tensorboard/webapp/feature_flag/data_source:feature_flag_testing",
         "//tensorboard/webapp/feature_flag/store",
         "//tensorboard/webapp/feature_flag/store:types",
         "@npm//@ngrx/store",
-        "@npm//@types/jasmine",
-    ],
-)
-
-tf_ng_module(
-    name = "feature_flag_types",
-    srcs = [
-        "tb_feature_flag_data_source_types.ts",
-    ],
-    deps = [
-        "//tensorboard/webapp/feature_flag:types",
-        "@npm//@angular/core",
-    ],
-)
-
-tf_ng_module(
-    name = "feature_flag",
-    srcs = [
-        "query_params.ts",
-        "tb_feature_flag_data_source.ts",
-        "tb_feature_flag_module.ts",
-    ],
-    deps = [
-        ":feature_flag_types",
-        "//tensorboard/webapp/feature_flag:types",
-        "//tensorboard/webapp/types",
-        "@npm//@angular/common",
-        "@npm//@angular/core",
-        "@npm//rxjs",
-    ],
-)
-
-tf_ng_module(
-    name = "feature_flag_testing",
-    testonly = True,
-    srcs = [
-        "tb_feature_flag_testing.ts",
-    ],
-    deps = [
-        ":feature_flag_types",
-        "//tensorboard/webapp/feature_flag:testing",
-        "//tensorboard/webapp/feature_flag:types",
-        "@npm//@angular/core",
-    ],
-)
-
-tf_ng_module(
-    name = "feature_flag_test_lib",
-    testonly = True,
-    srcs = [
-        "tb_feature_flag_data_source_test.ts",
-    ],
-    deps = [
-        ":feature_flag",
-        "//tensorboard/webapp/angular:expect_angular_core_testing",
-        "@npm//@angular/core",
         "@npm//@types/jasmine",
     ],
 )
